feat: refactor authentication routes and add auth.company route
